Split request spec helpers into multiple files.
