+++ spec/support/request_helpers/actions.rb
@@ -1,0 +1,7 @@
+module RequestHelpers
+  module Actions
+    def put_content_item(body: content_item.to_json)
+      put request_path, body
+    end
+  end
+end
